Move TripIdAndServiceDate to a top level class and convert to a record.
